requirements.txt
dotmap==1.3.30
pycryptodome==3.23.0
